important update - Enemy has IsNearEdge method
added a method to stop enemy from attacking the player near the edge of the ground layer and to stop moving. if player arrives from other side of enemy it will resume agro on player. enemy will not fall of ground when near the edge. enemy gismos shoot a red line with float 0.5f to detect IsGrounded